+++ src/main/kotlin/com/github/tsdaer/dreamshaderlanguagesupport/language/refactoring/DreamShaderRefactoringSupportProvider.kt
@@ -1,0 +1,15 @@
+package com.github.tsdaer.dreamshaderlanguagesupport.language.refactoring
+
+import com.github.tsdaer.dreamshaderlanguagesupport.language.psi.DreamShaderDeclaration
+import com.intellij.lang.refactoring.RefactoringSupportProvider
+import com.intellij.psi.PsiElement
+
+class DreamShaderRefactoringSupportProvider : RefactoringSupportProvider() {
+    override fun isSafeDeleteAvailable(element: PsiElement): Boolean {
+        return element is DreamShaderDeclaration && element.nameIdentifier != null
+    }
+
+    override fun isInplaceRenameAvailable(element: PsiElement, context: PsiElement?): Boolean {
+        return element is DreamShaderDeclaration && element.nameIdentifier != null
+    }
+}