+++ src/main/kotlin/com/github/tsdaer/dreamshaderlanguagesupport/language/refactoring/DreamShaderRenameInputValidator.kt
@@ -1,0 +1,18 @@
+package com.github.tsdaer.dreamshaderlanguagesupport.language.refactoring
+
+import com.github.tsdaer.dreamshaderlanguagesupport.language.psi.DreamShaderDeclaration
+import com.intellij.patterns.ElementPattern
+import com.intellij.patterns.PlatformPatterns
+import com.intellij.psi.PsiElement
+import com.intellij.refactoring.rename.RenameInputValidator
+import com.intellij.util.ProcessingContext
+
+class DreamShaderRenameInputValidator : RenameInputValidator {
+    override fun getPattern(): ElementPattern<out PsiElement> {
+        return PlatformPatterns.psiElement(DreamShaderDeclaration::class.java)
+    }
+
+    override fun isInputValid(newName: String, element: PsiElement, context: ProcessingContext): Boolean {
+        return DreamShaderNamesValidator().isIdentifier(newName, element.project)
+    }
+}