+++ src/test/kotlin/com/github/tsdaer/dreamshaderlanguagesupport/language/navigation/DreamShaderFindUsagesProviderTest.kt
@@ -23,7 +23,7 @@ class DreamShaderFindUsagesProviderTest : BasePlatformTestCase() {
 
         val provider = DreamShaderFindUsagesProvider()
         val typeText = provider.getType(declaration!!)
-        assertEquals("dreamshader shaderfunction declaration", typeText)
+        assertEquals("dreamshader ShaderFunction declaration", typeText)
     }
 
     fun testGetTypeFallsBackWhenKeywordMissing() {
@@ -33,5 +33,45 @@ class DreamShaderFindUsagesProviderTest : BasePlatformTestCase() {
         assertNotNull(token)
         assertEquals("dreamshader symbol", provider.getType(token!!))
     }
-}
 
+    fun testCanFindUsagesForNameAttributeValueToken() {
+        val file = myFixture.configureByText(
+            "find_usages_name_attribute_token.dsf",
+            """
+            ShaderFunction(Name="Functions/F_PulseTint")
+            {
+                Inputs = {
+                    vec3 InColor;
+                }
+            }
+            """.trimIndent()
+        )
+
+        val provider = DreamShaderFindUsagesProvider()
+        val nameAttributeValue = file.findElementAt(file.text.indexOf("\"Functions/F_PulseTint\"") + 1)
+        assertNotNull("Expected Name attribute string token", nameAttributeValue)
+        assertTrue(provider.canFindUsagesFor(nameAttributeValue!!))
+        assertEquals("F_PulseTint", provider.getDescriptiveName(nameAttributeValue))
+        assertEquals("ShaderFunction F_PulseTint", provider.getNodeText(nameAttributeValue, false))
+    }
+
+    fun testGetNodeTextUsesQualifiedNamespaceNameWhenRequested() {
+        val file = myFixture.configureByText(
+            "find_usages_qualified_node_text.dsh",
+            """
+            Namespace Tools {
+                Function Blend {
+                }
+            }
+            """.trimIndent()
+        )
+
+        val declarationOffset = file.text.indexOf("Function Blend") + "Function ".length
+        val declaration = file.findElementAt(declarationOffset)!!.parent as DreamShaderDeclaration
+        val provider = DreamShaderFindUsagesProvider()
+
+        assertEquals("Tools::Blend", provider.getDescriptiveName(declaration))
+        assertEquals("Function Blend", provider.getNodeText(declaration, false))
+        assertEquals("Function Tools::Blend", provider.getNodeText(declaration, true))
+    }
+}